Move map view to the fragment
