+++ app/build.gradle
@@ -4,8 +4,13 @@ apply plugin: 'kotlin-android'
 
 apply plugin: 'kotlin-android-extensions'
 
+apply plugin: 'kotlin-kapt'
+
 android {
 	compileSdkVersion 29
+	dataBinding {
+		enabled = true
+	}
 	buildToolsVersion "29.0.2"
 	defaultConfig {
 		applicationId "test.mug.espresso"
@@ -30,6 +35,8 @@ dependencies {
 	implementation 'androidx.core:core-ktx:1.0.2'
 	implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 	implementation 'com.google.android.gms:play-services-maps:16.1.0'
+	implementation 'androidx.navigation:navigation-fragment-ktx:2.1.0'
+	implementation 'androidx.navigation:navigation-ui-ktx:2.1.0'
 	testImplementation 'junit:junit:4.12'
 	androidTestImplementation 'androidx.test:runner:1.1.1'
 	androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
