+++ app/src/main/java/test/mug/espresso/MainActivity.kt
@@ -1,42 +1,13 @@
 package test.mug.espresso
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.OnMapReadyCallback
-import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.MarkerOptions
-
-class MainActivity : AppCompatActivity(), OnMapReadyCallback {
-
-	private lateinit var mMap: GoogleMap
+class MainActivity : AppCompatActivity() {
 
 	override fun onCreate(savedInstanceState: Bundle?) {
 		super.onCreate(savedInstanceState)
+
 		setContentView(R.layout.activity_main)
-		// Obtain the SupportMapFragment and get notified when the map is ready to be used.
-		val mapFragment = supportFragmentManager
-			.findFragmentById(R.id.map) as SupportMapFragment
-		mapFragment.getMapAsync(this)
-	}
-
-	/**
-	 * Manipulates the map once available.
-	 * This callback is triggered when the map is ready to be used.
-	 * This is where we can add markers or lines, add listeners or move the camera.
-	 * If Google Play services is not installed on the device, the user will be prompted to install
-	 * it inside the SupportMapFragment. This method will only be triggered once the user has
-	 * installed Google Play services and returned to the app.
-	 */
-	override fun onMapReady(googleMap: GoogleMap) {
-		mMap = googleMap
-
-		// Add a marker in Wroclaw and move the camera
-		val wroclaw = LatLng(51.1079, 17.0385)
-		mMap.addMarker(MarkerOptions().position(wroclaw).title("Marker in Wroclaw"))
-		mMap.animateCamera(CameraUpdateFactory.newLatLngZoom(wroclaw, 13f), 500, null)
 	}
 }
