+++ app/src/main/res/navigation/navigation.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:id="@+id/navigation"
+	app:startDestination="@id/mapViewFragment">
+
+	<fragment
+		android:id="@+id/mapViewFragment"
+		android:name="test.mug.espresso.MapViewFragment"
+		android:label="fragment_map_view"
+		tools:layout="@layout/fragment_map_view" />
+</navigation>
